Add GdTwoLinesRenderer
It's a custom text cell renderer that renders an optional second line
dim, with a maximum number of overall available lines.
If there's no text for the second line, all the space will be allocated
to the first, and ellipsized accordingly; otherwise, num_lines -1 will
be allocated to the main text and the last line to the dim label.