all: implement refresh information for single elements
This way we can be notified live of the changes in the model on one of
the displayed resources.
The code seems to hit a weird bug where toggling the favorite tag only
works once; still to investigate why, but the wrong results seem to come
directly from the TrackerSparqlCursor.