loader: try to load gdocs from the last saved cache as a fallback
E.g. if we have no network, we still want to see the last saved cache,
if any. To do this, we need to change the saved name of the cache to
hash the document-id instead of the entry id, since this last one could
have multiple valid values for each document entity.