all: implement listening to create/delete events from the tracker DB
We now associate a type for each change, and add/remove elements from
the view if we detect such a change in the graph (and the changed
elements match what we're interested in).

This way of ensuring change events are always propagated to the view
allows us to remove a refresh of the view after the GData miner has been
spawned (as we will receive the changes through the GraphUpdated
signal and update single elements on the view) but to work properly this
will need changes in the GData miner itself, which for instance, doesn't
track deletes right now.