query: apply filters as usual when doing a single element query
Single element queries are done in respose to changes in the tracker DB,
both when an element changes or when a new one is added.
We need to use the same filters we would use for an "overview" query in
this case too, otherwise a document not matching the current filters
could be added to the filtered view.

We also need to move the filter-by-type function from WHERE to
FILTER, as UNION doesn't seem to work with a specified resource value.