sidebar: remove the sidebar
As for the latest iterations of the designs, we now want to be
sidebar-less.
The sidebar currently controls categories and collections: collections
will come back in the next commit as regular view items, while
categories for now have been added to the search bar dropdown, waiting
for a better place.