searchbar: Split search into words
Rather than using the search entry text as a single term when building
the query, split the text into multiple terms and concatenate the
results. That way, a search can be refined by adding more words to a
query, which is generally more useful (possibly with the exception of
full-text search).

https://bugzilla.gnome.org/show_bug.cgi?id=665158