searchbar: integrate with GdTaggedEntry
Add and remove tags for each of the categories we filter from the
dropdown.
Use the close button on the tag to remove the filter; clicking in
another part of the tag triggers the dropdown menu again, if it's not in
already.
Additionally, close the dropdown menu as soon as an item is chosen,
since now we can more easily tweak search parameters using tags.