all: add GdMainView/GdMainIconView/GdMainListView
Instead of implementing the view classes in JS, for the sake of sharing
them with other Findind and Reminding applications, split the main view
into an interface (GdMainView) with two concrete implementations for
Icon/List views.
An additional GdMainViewWidget will eventually take the place of the
View JS base class.