lib: add GdTogglePixbufRenderer
This is a pixbuf cell renderer with an active state on top of it; it
uses gtk_render_check() to render a themable checkbutton image.
Unfortunately we have to use a style property for the image size, since
we don't have support for that as a CSS property yet. GdMainIconView
will add a style property in a subsequent commit.