sidebar-thumbnails: don't set icon view columns to G_MAXINT
GTK+ behavior wrt. this has changed since commit
d406bf96d4bb1191b05f1ddb374c2f57138c8d45, but it was a hack in the first
place. GTK+ shouldn't probably crash if we do that anyway, but in the
meantime, set this value to the number of items of the document
displayed.